+++ src/main/java/com/attendanceAPI/models/TimeTable.java
@@ -24,7 +24,7 @@ public class TimeTable  {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	
 	@OneToMany( cascade = CascadeType.ALL)
 	@JoinTable(name="timeTableSlot", joinColumns=
@@ -36,11 +36,11 @@ public class TimeTable  {
 	@OneToOne(cascade=CascadeType.ALL)
 	private Lecture lecture; 
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -64,8 +64,7 @@ public class TimeTable  {
 	@Override
     public String toString() {
         return new Gson().toJson(this);
-    }
-	
+    }
