+++ src/main/java/com/attendanceAPI/repository/TimeTableRepository.java
@@ -1,6 +1,8 @@
 package com.attendanceAPI.repository;
 
 
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +10,9 @@ import org.springframework.stereotype.Repository;
 import com.attendanceAPI.models.TimeTable;
 
 @Repository
-public interface TimeTableRepository extends JpaRepository<TimeTable, Long> {
+public interface TimeTableRepository extends JpaRepository<TimeTable, Integer> {
+
+
