+++ src/main/java/com/attendanceAPI/service/TimeTableService.java
@@ -48,7 +48,7 @@ public class TimeTableService {
 		}
 	}
 	
-	public Optional<TimeTable> findLectureById(long id)
+	public Optional<TimeTable> findLectureById(Integer id)
 	{
 		try {
